pyproject.toml

[tool.poetry]
name = "navidificador"
version = "0.0.1"
description = ""
authors = [
  "Vicent Alabau <valabau@sciling.com>"
]
packages = [
    { include = "navidificador", from = "src" },
    { include = "bin", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
sentry-sdk = "^1.10.1"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
python-multipart = "^0.0.5"
python-magic = "^0.4.27"
pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
autopep8 = "==1.5.7"
bandit = "==1.7.0"
black = "==22.3.0"
coverage = "==5.5"
dependency-check = "==0.6.0"
flake8 = "==3.9.2"
isort = "==5.10.1"
ipykernel = "5.5.6"
liccheck = "==0.6.5"
moto = "==2.2.16"
mypy = "==0.961"
pylint = "==2.15.5"
pytest = "==6.2.4"
pytest-cov = "==2.12.1"
pytest-mock = "==3.6.1"
pytest-reraise = "==2.1.1"
pytest-xdist = "==2.3.0"
safety = "==1.10.3"
pre-commit = "==2.20.0"
packaging = "==21.3"
types-python-dateutil = "==2.8.17"
python-sonarqube-api = "==1.3.0"
"ruamel.yaml" = "==0.17.21"
toml = "==0.10.2"
coloredlogs = "==15.0.1"
prospector = "^1.7.7"

[tool.poetry.scripts]
start = "bin.server:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
skip_glob = ["/setup.py"]
filter_files = true


[tool.black]
line-length = 160
include = '.pyi?$'


[tool.coverage.run]
omit = [
  ".venv/*",
  "tests/*",
  "**/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=. --cov-report xml:.reports/coverage.xml --cov-report html:.reports/htmlcov  --verbose --capture=no -p no:sugar"
testpaths = [
    "tests",
]


[tool.bandit]
exclude_dirs = [".venv"]


[tool.liccheck]
# Authorized and unauthorized licenses in LOWER CASE
authorized_licenses = [
    "bsd",
    "new bsd",
    "bsd license",
    "new bsd license",
    "simplified bsd",
    "3-clause bsd",
    "BSD-3-Clause",
    "apache",
    "apache 2.0",
    "apache license 2.0",
    "apache software license",
    "asl",
    "asl 2",
    "apache-software",
    "apache software",
    "gnu lgpl",
    "lgpl",
    "lgpl with exceptions or zpl",
    "GNU Library or Lesser General Public License (LGPL)",
    "lgplv2",
    "lgplv2+",
    "GNU Lesser General Public License v2 or later (LGPLv2+)",
    "isc license",
    "isc license (iscl)",
    "mit",
    "mit license",
    "mozilla public license 1.1 (mpl 1.1)",
    "mozilla public license 2.0 (mpl 2.0)",
    "mpl-2.0",
    "python software foundation license",
    "python software foundation",
    "psf",
    "public domain <http://unlicense.org>",
    "zpl 2.1",
]

unauthorized_licenses = [
    "gpl v2",
    "gpl v3",
    "gpl version 2",
]
